+++ rmp/tests/func/decode/mod.rs
@@ -1,15 +1,10 @@
-// Low-level interface.
 mod null;
 mod bool;
 mod sint;
 mod uint;
 mod float;
-// mod string;
-// mod bin;
-// mod array;
-// mod map;
-// mod ext;
-
-// // High-level deserialize interface.
-// mod value;
-// mod value_ref;
+mod string;
+mod bin;
+mod array;
+mod map;
+mod ext;
