+++ rmp/tests/func/encode/mod.rs
@@ -1,4 +1,3 @@
-// Low-level interface.
 mod null;
 mod bool;
 mod int;
@@ -8,7 +7,3 @@ mod bin;
 mod array;
 mod map;
 mod ext;
-
-// High-level serialize interface.
-// mod value;
-// mod value_ref;
